schemas structure reorg
